requirements.txt
coloredlogs==15.0.1
atlassian-python-api==3.41.16
pytz==2024.2
